requirements.txt
google-generativeai
python-dotenv
flask
sqlite3